role permission manage and use for user crud - admin panel

+++ app/Http/Controllers/Admin/User/DeleteController.php
@@ -10,6 +10,9 @@ class DeleteController extends Controller
 {
     public function delete(User $user)
     {
+        if(!auth()->user()->hasPermissionTo('delete-any-user'))
+            abort(403);
+
         $user->delete();
         return back()->with('success','کاربر مورد نظر با موفقیت حذف شد.');
     }
